Fix: Change reportEndCallWithUUID reason parameter from int to double
- TurboModule codegen translates TypeScript number to Java double
- Updated implementation and oldarch spec to use double instead of int
- Fixes compilation error in new architecture builds

+++ android/src/main/java/com/callkeeper/CallKeeperModule.java
@@ -262,7 +262,7 @@ public class CallKeeperModule extends CallKeeperSpec {
     }
 
     @ReactMethod
-    public void reportEndCallWithUUID(String callUUID, int reason, Promise promise) {
+    public void reportEndCallWithUUID(String callUUID, double reason, Promise promise) {
         try {
             VoiceConnectionService.endCall(callUUID);
             calls.remove(callUUID);
